Renaming of features and documentation
