+++ Cargo.toml
@@ -3,7 +3,7 @@ name = "prometheus_push"
 version = "0.2.1"
 edition = "2021"
 authors = ["Mathias Oertel <mathias.oertel@pm.me>"]
-description = "Crate to extends the prometheus crate with async pushgateway support"
+description = "Crate to extends prometheus crates with pushgateway support"
 documentation = "https://docs.rs/prometheus_push"
 readme = "README.md"
 keywords = ["prometheus", "pushgateway", "metrics", "oberservability"]
@@ -21,9 +21,9 @@ reqwest = { version = "0.11", default-features = false, optional = true }
 log = { version = "0.4", default-features = false, optional = true }
 
 [features]
-default = ["async"]
-async = ["async-trait"]
+default = ["non_blocking"]
+non_blocking = ["async-trait"]
 blocking = []
 prometheus_crate = ["prometheus"]
-with_reqwest = ["async", "reqwest", "log"]
+with_reqwest = ["non_blocking", "reqwest", "log"]
 with_reqwest_blocking = ["blocking", "reqwest/blocking", "log"]
