+++ README.md
@@ -1,7 +1,8 @@
 # Prometheus push
 
-This crate works as an extension to the [prometheus](https://crates.io/crates/prometheus) crate to be able to push non-blocking (default) to your
-Prometheus pushgateway and with a less dependent setup of `reqwest` (no `openssl` for example) or with an implementation of your own http client.
+This crate works as an extension to prometheus crates like [prometheus](https://crates.io/crates/prometheus) to be able to push non-blocking (default)
+or blocking to your Prometheus pushgateway and with a less dependent setup of `reqwest` (no `openssl` for example) or with an implementation of your
+own http client.
 
 By default you have to implement the `Push` trait to use it with your choice of http client or you can use the `with_reqwest` feature.
 This feature already implements `Push` in a `PushClient` that leverages `reqwest` under the hood. Reqwest is setup without default features
@@ -11,17 +12,28 @@ Async functionality is considered the standard in this crate but you can enable 
 can enable the corresponding blocking `reqwest` implementation with the `with_reqwest_blocking` feature in which case you enable the `blocking`
 feature of the `reqwest` crate as well.
 
-## Example with feature `with_reqwest`
+In terms of the underlying prometheus functionality you have to implement the `ConvertMetrics` trait or you use the already implemented feature
+`prometheus_crate` that leverages the [prometheus](https://crates.io/crates/prometheus) crate.
+
+## Example with features `with_reqwest` and `prometheus_crate`
 
 ```rust
+use prometheus::core::Collector;
 use prometheus::labels;
+use prometheus::proto::MetricFamily;
+use prometheus_push::non_blocking::MetricsPusher;
+use prometheus_push::prometheus_crate::PrometheusMetricsConverter;
 use prometheus_push::with_reqwest::PushClient;
 use prometheus_push::MetricsPusher;
 use reqwest::Client;
 use url::Url;
 
+pub type PrometheusMetricsPusher =
+  MetricsPusher<PushClient, PrometheusMetricsConverter, MetricFamily, Box<dyn Collector>>;
+
 let push_gateway: Url = <address to your instance>;
-let metrics_pusher = MetricsPusher::<PushClient>::from(Client::new(), &push_gateway)?;
+let client = Client::new();
+let metrics_pusher = PrometheusMetricsPusher::from(client, &push_gateway)?;
 metrics_pusher
   .push_all(
     "<your push jobs name>",
@@ -58,12 +70,40 @@ impl Push for YourClient {
 }
 ```
 
+## Implement `ConvertMetrics` yourself
+
+In case you want to use another promethues client implementation you can implement your own type that implements
+the `ConvertMetrics` trait to inject it into your instance of `MetricsPusher`.
+
+```rust
+impl ConvertMetrics<YourMetricFamily, Box<dyn YourCollector>> for YourMetricsConverter {
+    fn metric_families_from(
+        &self,
+        collectors: Vec<Box<dyn YourCollector>>,
+    ) -> Result<Vec<YourMetricFamily>> {
+        // implement the conversion from your Collectors to your MetricsFamilies 
+    }
+
+    fn create_push_details(
+        &self,
+        job: &str,
+        url: &Url,
+        grouping: &HashMap<&str, &str>,
+        metric_families: Vec<YourMetricFamily>,
+    ) -> Result<(Url, Vec<u8>, String)> {
+        // create your push details for the `Push` methods: Url, body and content type
+    }
+}
+```
+
 ## Features
 
 - `default`: by default async functionality and no reqwest is enabled
-- `with_reqwest`: this feature enables `reqwest` in minimal configuration and enables the alredy implemented `PushClient`
+- `non_blocking`: this ennables the async functionality
 - `blocking`: on top of the default feature you get the same functionality in a blocking fashion
-- `with_reqwest_blocking`: like `with_reqwest` but including blocking and completely blocking
+- `with_reqwest`: this feature enables the `non_blocking` feature as well as `reqwest` in minimal configuration and enables the alredy implemented `PushClient`
+- `with_reqwest_blocking`: like `with_reqwest` but including `blocking` instead of `non_blocking`
+- `prometheus_crate`: enables the functionality of the [prometheus](https://crates.io/crates/prometheus) crate
 
 ## License
